==> R/03_augmented.qmd <==
---
title: "Data augmentation"
format:
  html:
    embed-resources: true
editor: visual
---

## Load libraries

```{r}
#| warning: false
#| message: false

library("tidyverse")
source("99_proj_func.R")
```

## Load data

```{r}
#| warning: false
#| message: false

metadata_clean <- read_tsv("../data/02_dat_metadata_clean.tsv")
raw_counts <- read_tsv("../data/01_dat_raw_counts.tsv")
```

## Augment metadata, normalize counts and join tables

```{r}
# Creating age groups
metadata_clean <- metadata_clean |> 
  mutate(Age_groups = cut(Age,
                          breaks = c(18,40,60,82),
                          labels = c("Young", "Middle", "Older"),
                          right = FALSE))

# Normalizing counts
cpm_counts <- CPM_normalization(raw_counts)
quant_counts <- quantile_normalization(raw_counts)

# Calculating GEP scores
augmented_metadata <- GEP_score_calculation(quant_counts, 
                                            metadata_clean)

augmented_metadata <- augmented_metadata |>
  mutate(TLS_status_bin = if_else(TLS_status == "control", 
                                  0, 1))

# Pivoting normalized counts to long format 
cpm_long <- cpm_counts |> 
  pivot_longer(cols = -GeneFeature,
               names_to = "Sample_names",
               values_to = "CPM")
quant_long <- quant_counts |> 
  pivot_longer(cols = -GeneFeature,
               names_to = "Sample_names",
               values_to = "Quantile")

# Joining data
joined_data <- quant_long |> 
  left_join(cpm_long, 
            by = c("Sample_names", "GeneFeature")) |> 
  left_join(augmented_metadata, 
            by = "Sample_names")

# Binarizing TLS_status and log-transforming CPM counts
joined_data <- joined_data |>
  mutate(CPM_log2 = log(CPM+1, base=2))

keep_genes_for_PCA <- cpm_counts |>
  filter(rowSums(across(!GeneFeature)) > 0) |>
  select(GeneFeature) |>
  pull()

PCA_ready_table <- joined_data |>
  select(GeneFeature, 
         CPM, 
         Sample_names) |>
  filter(GeneFeature %in% keep_genes_for_PCA) |>
  pivot_wider(names_from = GeneFeature,
              values_from = CPM) |>
  inner_join(augmented_metadata, 
             by = "Sample_names")

# Saving final table
write_tsv(joined_data, 
          file = "../data/03_augmented_table.tsv")

write_tsv(PCA_ready_table, 
          file = "../data/03_augmented_table_PCA.tsv")

write_tsv(augmented_metadata,
          file = "../data/03_augmented_metadata.tsv")
```
